Split commands/revert.rs into plan and apply submodules
revert/mod.rs keeps the plan/apply entry points, the revert plan data
model, and shared node/path helpers; plan.rs builds per-repo revert
plans from ledger nodes; apply.rs owns preflight and execution for both
local revert/cherry-pick commits and provider PR reverts. Pure moves,
no behavior change.

Knit-Group: kg_20260612_f5d5a1
Knit-Bundle: open-source-prep
